Create pt_pt.json
Language pt_PT (Portuguese Portugal) completely added, i am an adult native speaker (born and always have lived here) with a lot of programming experience, and I think i can say that my portuguese and english are both close to perfect (specially portuguese, of course). Portuguese from Portugal is really complex (unlike brasilian portuguese, which by the way, we don't speak it! We undestand it, but we do not speak it and it's awkward for us to have something in pt_br, because in a lot of things they use really awkward words for some things... Welp, just completely different dialects, but long story short, we over here really hate to use pt_br), and so because it is complex, it's usually only well done by a native speaker. I did my best, we have a server with this plugin and we really love it, so i'm looking forward to have this native PT added!